split out utility functions and changed babel to transpile both files into a concatenated file
