--- index.js
@@ -1,202 +1,0 @@
-const fs = require('fs');
-const path = require('path');
-
-module.exports = PathWizardModule;
-
-if (require.cache && __filename) delete require.cache[__filename];
-
-function PathWizardModule(target = module.parent.require, rootPath, options) {
-  if (rootPath && typeof rootPath !== 'string') throw new Error('PathWizard constructor only accepts undefined or a string-typed project directory.');
-  const _PathWizard = new PathWizard(rootPath, options);
-  return _PathWizard.proxy(target);
-}
-
-function PathWizard(rootPath = process.cwd(), { cache = true, ignored = ['node_modules', 'bower_components'] } = {}) {
-  this.root = rootPath;
-  this.ignored = ignored;
-  this.cache = !!cache;
-  this.nodes = this.cache ? traverse(this.root, this.ignored) : [];
-}
-
-PathWizard.prototype.abs = function (filePath) {
-  if (!filePath) throw new Error(`A search expression must be provided to the 'abs' method.`);
-  if (!filePath.length) throw new Error(`The 'abs' method requires a non-empty string.`);
-
-  let _filePath, _filePathWithIndex, matches;
-  if (filePath === '/') {
-    _filePath = ['index.js'];
-  } else {
-    _filePath = Array.isArray(filePath) ? filePath : filePath.split(path.sep);
-    if (_filePath[_filePath.length - 1] === '') _filePath.pop();
-    if (_filePath[0] === '.' || _filePath[0] === '') _filePath[0] = '~';
-  }
-
-  if (!this.cache) {
-    this.nodes = traverse(this.root, this.ignored);
-  }
-
-  let target = _filePath[_filePath.length - 1];
-  if (target.indexOf('.') < 0) {
-    _filePath.push(`${_filePath.pop()}.js`);
-    _filePathWithIndex = _filePath.slice();
-    _filePathWithIndex.push(_filePathWithIndex.pop()
-      .replace(/\.\w+/, ''), 'index.js');
-  }
-
-  matches = findMatchingDirectories.bind(this)(_filePath);
-  if (!matches.length && _filePathWithIndex)
-    matches = findMatchingDirectories.bind(this)(_filePathWithIndex);
-  if (matches.length === 1) return path.join(this.root, ...matches.pop().slice(1));
-  else err.bind(this)(filePath, matches);
-}
-
-PathWizard.prototype.rel = function (filePath) {
-  if (!filePath) throw new Error(`A search expression must be provided to the 'abs' method.`);
-  if (!filePath.length) throw new Error(`The 'abs' method requires a non-empty string.`);
-
-  const _to = path.normalize(this.abs(filePath));
-  const _from = module.parent.filename;
-  const rel = path.relative(_from, _to).slice(3);
-
-  return /\.\./.test(rel) ? rel : `./${rel}`;
-}
-
-PathWizard.prototype.ignore = function (expressions) {
-  ignorePath(expressions, this.ignored);
-  return this;
-};
-
-PathWizard.prototype.absDir = function (filePath) {
-  if (!filePath) throw new Error(`A search expression must be provided to the 'abs' method.`);
-  if (!filePath.length) throw new Error(`The 'abs' method requires a non-empty string.`);
-
-  let _filePath, matches;
-  if (filePath === '/') {
-    return path.normalize(this.root);
-  } else {
-    _filePath = Array.isArray(filePath) ? filePath : filePath.split(path.sep);
-    if (_filePath[_filePath.length - 1] === '') _filePath.pop();
-    if (_filePath[0] === '.' || _filePath[0] === '') _filePath[0] = '~';
-  }
-  if (this.cache && !this.nodes.length) {
-    this.nodes = traverse(this.root, this.ignored);
-  } else if (!this.cache) {
-    this.nodes = traverse(this.root, this.ignored);
-  }
-
-  matches = findMatchingDirectories.bind(this)(_filePath);
-  if (matches.length === 1) return path.join(this.root, ...matches.pop().slice(1));
-  else err.bind(this)(filePath, matches);
-};
-
-PathWizard.prototype.relDir = function (filePath) {
-  if (!filePath) throw new Error(`A search expression must be provided to the 'abs' method.`);
-  if (!filePath.length) throw new Error(`The 'abs' method requires a non-empty string.`);
-
-  const to = path.normalize(this.absDir(filePath));
-  const rel = path.relative('', to);
-
-  return /\.\./.test(rel) ? rel : `./${rel}`;
-};
-
-PathWizard.prototype.proxy = function (target) {
-  return new Proxy(target, {
-    apply: (target, thisArg, argumentList) => req(target, this.abs.bind(this), ...argumentList),
-    get: (target, property) => {
-      switch (property) {
-        case 'abs':
-          return this.abs.bind(this);
-        case 'absDir':
-          return this.absDir.bind(this);
-        case 'rel':
-          return this.rel.bind(this);
-        case 'relDir':
-          return this.relDir.bind(this);
-        case 'ignore':
-          return this.ignore.bind(this);
-        case 'root':
-          return this.root;
-        case 'nodes':
-          return this.nodes;
-        case 'cache':
-          return this.cache;
-        case 'ignored':
-          return this.ignored;
-        default:
-          return target[property];
-      }
-    }
-  })
-}
-
-function traverse(rootPath = process.cwd(), ignoredArray, directory = '', nodesArray = []) {
-  const nodes = fs.readdirSync(path.join(rootPath, directory))
-    .filter(n => {
-      const name = n.split(path.sep)
-        .pop();
-      return !isPathIgnored(name, ignoredArray);
-    });
-
-  nodes.forEach(node => {
-    if (fs.statSync(path.join(rootPath, directory, node)).isDirectory()) {
-      traverse(rootPath, ignoredArray, path.join(directory, node), nodesArray);
-    }
-    nodesArray.push(path.join(directory, node)
-      .split(path.sep));
-  });
-  return nodesArray.map(prependRoot);
-}
-
-function req(target, findingFunction, filePath) {
-  if (!filePath) throw new Error(`A search expression must be provided to the 'req' method.`);
-  if (!filePath.length) throw new Error(`The 'req' method requires a non-empty string or array search expression.`);
-
-  let mod;
-  try {
-    mod = target(filePath);
-  } catch (e) {
-    mod = target(findingFunction(filePath));
-  }
-  return mod;
-};
-
-function findMatchingDirectories(_filePath) {
-  const matches = [];
-  this.nodes.forEach(node => {
-    const _path = _filePath.slice();
-    const _node = node.slice();
-    while (_path.length) {
-      if (_path.pop() !== _node.pop()) return;
-    }
-    matches.push(node);
-  })
-  return matches;
-}
-
-function isPathIgnored(pathSegment, ignoredArray) {
-  if (pathSegment[0] === '.') return true;
-  return ignoredArray.some(element => element === pathSegment);
-}
-
-function ignorePath(pathSegment, ignored) {
-  if (Array.isArray(pathSegment)) {
-    pathSegment.forEach(exp => {
-      if (typeof exp !== 'string') `Ignored files and directories must be strings.${'\n'}`;
-    })
-    ignored.push(...pathSegment);
-  }
-  ignored.push(pathSegment);
-}
-
-function prependRoot(node) {
-  if (node[0] !== '~') node.unshift('~');
-  return node;
-}
-
-function err(filePath, matches) {
-  if (!matches.length)
-    throw `No files in ${this.root} matched ${filePath}${'\n'}`;
-  else
-    throw `The path did not uniquely resolve! ${'\n\n'}${matches.map (match => path.join(...match)).join('\n')}${'\n'}`;
-}
-
