+++ package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "test": "mocha tests/index.js",
     "test-travis": "npm run transpile && istanbul cover ./node_modules/mocha/bin/_mocha --report lcovonly -- tests",
-    "transpile": "babel index.js --out-file src/index.js --presets=es2015"
+    "transpile": "babel src --out-file dist/index.js --presets=es2015"
   },
   "devDependencies": {
     "babel-cli": "^6.18.0",
@@ -25,3 +25,4 @@
     "mocha-lcov-reporter": "^1.2.0"
   }
 }
+
