+++ tests/index.js
@@ -2,7 +2,7 @@ const chai = require('chai');
 const spies = require('chai-spies');
 const fse = require('fs-extra');
 const path = require('path');
-const PathWizard = require('../src');
+const PathWizard = require('../dist');
 const expect = chai.expect;
 
 chai.use(spies);
